# openapi (2).json
{
  "openapi": "3.1.1",
  "info": {
    "title": "alp-pay.com API",
    "description": "Gateway API\n",
    "license": {
      "name": "Apache 2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.100"
  },
  "servers": [
    {
      "url": "https://engine-sandbox.alp-pay.com",
      "description": "Sandbox"
    },
    {
      "url": "https://engine.alp-pay.com",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/v1/payments": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Get payments",
        "description": "Get a list of payments sorted by creation date (most recent first)",
        "operationId": "getPayments",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "The number of items to skip before starting to collect the result set. Default is 0.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The numbers of items to return. Default is 50.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "created.gte",
            "in": "query",
            "description": "If passed, return only payments created at or after the specified time",
            "required": false,
            "schema": {
              "type": "string",
              "format": "ISO 8601 (YYYY-MM-DD'T'HH24:MI:SS)",
              "example": "2021-10-13T10:26:18"
            }
          },
          {
            "name": "created.lt",
            "in": "query",
            "description": "If passed, return only payments created strictly before the specified time",
            "required": false,
            "schema": {
              "type": "string",
              "format": "ISO 8601 (YYYY-MM-DD'T'HH24:MI:SS)",
              "example": "2021-10-13T10:39:34"
            }
          },
          {
            "name": "updated.gte",
            "in": "query",
            "description": "If passed, return only payments updated at or after the specified time",
            "required": false,
            "schema": {
              "type": "string",
              "format": "ISO 8601 (YYYY-MM-DD'T'HH24:MI:SS)",
              "example": "2021-10-13T10:26:18"
            }
          },
          {
            "name": "updated.lt",
            "in": "query",
            "description": "If passed, return only payments updated strictly before the specified time",
            "required": false,
            "schema": {
              "type": "string",
              "format": "ISO 8601 (YYYY-MM-DD'T'HH24:MI:SS)",
              "example": "2021-10-13T10:39:34"
            }
          },
          {
            "name": "referenceId.eq",
            "in": "query",
            "description": "If passed, return only payments with the requested referenceId",
            "required": false,
            "schema": {
              "type": "string",
              "example": "order-12345"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment List Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create payment",
        "description": "Payment request, used for DEPOSITS, WITHDRAWALS and REFUNDS",
        "operationId": "createPayment",
        "requestBody": {
          "description": "Payment Request - Request to initiate a payment",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequest"
              },
              "examples": {
                "default_example": {
                  "$ref": "#/components/examples/SimpleDepositRequest"
                },
                "basic_card_hpp_deposit_example": {
                  "$ref": "#/components/examples/BasicCardHppDepositRequest"
                },
                "basic_card_sts_deposit_example": {
                  "$ref": "#/components/examples/BasicCardStsDepositRequest"
                },
                "banktransfer_ar_deposit_example": {
                  "$ref": "#/components/examples/BankTransferArDepositRequest"
                },
                "pix_deposit_example": {
                  "$ref": "#/components/examples/PixDepositRequest"
                },
                "banktransfer_cl_deposit_example": {
                  "$ref": "#/components/examples/BankTransferClDepositRequest"
                },
                "banktransfer_cr_deposit_example": {
                  "$ref": "#/components/examples/BankTransferCrDepositRequest"
                },
                "banktransfer_gt_deposit_example": {
                  "$ref": "#/components/examples/BankTransferGtDepositRequest"
                },
                "banktransfer_hn_deposit_example": {
                  "$ref": "#/components/examples/BankTransferHnDepositRequest"
                },
                "banktransfer_pa_deposit_example": {
                  "$ref": "#/components/examples/BankTransferPaDepositRequest"
                },
                "banktransfer_uy_deposit_example": {
                  "$ref": "#/components/examples/BankTransferUyDepositRequest"
                },
                "crypto_btc_deposit_example": {
                  "$ref": "#/components/examples/CryptoBtcDepositRequest"
                },
                "crypto_btc_bep20_deposit_example": {
                  "$ref": "#/components/examples/CryptoBtcBep20DepositRequest"
                },
                "crypto_usdt_trc20_deposit_example": {
                  "$ref": "#/components/examples/CryptoUsdtTrc20DepositRequest"
                },
                "fiat_to_crypto_deposit_example": {
                  "$ref": "#/components/examples/FiatToCryptoDepositRequest"
                },
                "efecty_deposit_example": {
                  "$ref": "#/components/examples/EfectyDepositRequest"
                },
                "nequi_deposit_example": {
                  "$ref": "#/components/examples/NequiDepositRequest"
                },
                "pse_deposit_example": {
                  "$ref": "#/components/examples/PseDepositRequest"
                },
                "paytm_deposit_example": {
                  "$ref": "#/components/examples/PaytmDepositRequest"
                },
                "phonepe_deposit_example": {
                  "$ref": "#/components/examples/PhonePeDepositRequest"
                },
                "upi_collect_deposit_example": {
                  "$ref": "#/components/examples/UpiCollectDepositRequest"
                },
                "upi_intent_deposit_example": {
                  "$ref": "#/components/examples/UpiIntentDepositRequest"
                },
                "spei_deposit_example": {
                  "$ref": "#/components/examples/SpeiDepositRequest"
                },
                "banktransfer_pe_deposit_example": {
                  "$ref": "#/components/examples/BankTransferPeDepositRequest"
                },
                "paypo_deposit_example": {
                  "$ref": "#/components/examples/PayPoDepositRequest"
                },
                "basic_card_withdrawal_example": {
                  "$ref": "#/components/examples/BasicCardWithdrawalRequest"
                },
                "banktransfer_ar_withdrawal_example": {
                  "$ref": "#/components/examples/BankTransferArWithdrawalRequest"
                },
                "pix_withdrawal_example": {
                  "$ref": "#/components/examples/PixWithdrawalRequest"
                },
                "banktransfer_cl_withdrawal_example": {
                  "$ref": "#/components/examples/BankTransferClWithdrawalRequest"
                },
                "banktransfer_co_withdrawal_example": {
                  "$ref": "#/components/examples/BankTransferCoWithdrawalRequest"
                },
                "imps_withdrawal_example": {
                  "$ref": "#/components/examples/ImpsWithdrawalRequest"
                },
                "banktransfer_mx_withdrawal_example": {
                  "$ref": "#/components/examples/BankTransferMxWithdrawalRequest"
                },
                "banktransfer_pe_withdrawal_example": {
                  "$ref": "#/components/examples/BankTransferPeWithdrawalRequest"
                },
                "refund_example": {
                  "$ref": "#/components/examples/RefundRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Payment response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                },
                "examples": {
                  "deposit_example": {
                    "$ref": "#/components/examples/CreateDepositResponse"
                  },
                  "withdrawal_example": {
                    "$ref": "#/components/examples/CreateWithdrawalResponse"
                  },
                  "refund_example": {
                    "$ref": "#/components/examples/CreateRefundResponse"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "PaymentRequest": {
        "required": [
          "currency",
          "paymentType"
        ],
        "type": "object",
        "properties": {
          "referenceId": {
            "maxLength": 256,
            "type": "string",
            "description": "Reference assigned by Merchant. Will not go outside the system. Will be sent unchanged in the PaymentResponse.",
            "example": "payment_id=123;custom_ref=456"
          },
          "paymentType": {
            "$ref": "#/components/schemas/PaymentType"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "amount": {
            "multipleOf": 1e-18,
            "maximum": 1000000000,
            "minimum": 0.00001,
            "type": "number",
            "description": "Payment amount",
            "example": 11.12
          },
          "currency": {
            "type": "string",
            "description": "Payment currency",
            "format": "ISO 4217 code for FIAT currencies or cryptocurrency symbol",
            "example": "EUR"
          },
          "parentPaymentId": {
            "maxLength": 32,
            "type": "string",
            "description": "Id of initial deposit for refunds, Id of initial recurring payment for subsequent payments",
            "example": "91d27876e87f4b22b3ecd53924bf973d"
          },
          "description": {
            "maxLength": 512,
            "type": "string",
            "description": "Description of the transaction shown to the Customer. Can be sent outside the system.",
            "example": "Deposit 123 via TEST shop"
          },
          "card": {
            "$ref": "#/components/schemas/Card"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "billingAddress": {
            "$ref": "#/components/schemas/BillingAddress"
          },
          "returnUrl": {
            "maxLength": 512,
            "type": "string",
            "description": "URL to redirect Customer after processing",
            "example": "https://mywebsite.com/{id}/{referenceId}/{state}/{type}"
          },
          "webhookUrl": {
            "maxLength": 512,
            "type": "string",
            "description": "Url to receive payment status notifications",
            "example": "https://mywebsite.com/webhooks"
          },
          "startRecurring": {
            "type": "boolean",
            "description": "Send 'true' if you want this payment to initiate recurring chain. Default is 'false'."
          },
          "preAuth": {
            "type": "boolean",
            "description": "Send 'true' if you want to request a Pre-Authorization charge (initiate a 2-phase deposit)."
          },
          "recurringToken": {
            "type": "string",
            "description": "To continue recurring chain, send a token from a previously initiated recurring payment."
          },
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionRequest"
          },
          "additionalParameters": {
            "type": "object",
            "description": "Additional parameters required by some payment providers. Contact support for more information.",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "bankCode": "ABHY0065032",
              "countryOfBirth": "CY"
            }
          }
        }
      },
...
      "PaymentListResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "ISO 8601",
            "example": "2020-10-07T13:36:32.595+00:00"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 200
          },
          "hasMore": {
            "type": "boolean",
            "description": "Indicates if there are more pages to return"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Payment"
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BadRequestError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UnauthorizedError"
            }
          }
        }
      },
...
    },
    "examples": {
...
      "SimpleDepositRequest": {
        "summary": "Simple deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "currency": "EUR"
        }
      },
      "BasicCardHppDepositRequest": {
        "summary": "Basic card HPP deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "amount": 10.01,
          "currency": "GBP",
          "customer": {
            "email": "jw@mail.com",
            "phone": "44 02072243688"
          },
          "billingAddress": {
            "countryCode": "GB",
            "city": "London",
            "addressLine1": "221b Baker St, Marylebone",
            "postalCode": "NW1 6XE"
          }
        }
      },
      "BasicCardStsDepositRequest": {
        "summary": "Basic card STS deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "amount": 10.01,
          "currency": "GBP",
          "customer": {
            "email": "jw@mail.com",
            "phone": "44 02072243688"
          },
          "billingAddress": {
            "countryCode": "GB",
            "city": "London",
            "addressLine1": "221b Baker St, Marylebone",
            "postalCode": "NW1 6XE"
          },
          "card": {
            "cardNumber": "4000000000000002",
            "cardholderName": "John Watson",
            "cardSecurityCode": "123",
            "expiryMonth": "01",
            "expiryYear": "2030"
          }
        }
      },
      "BankTransferArDepositRequest": {
        "summary": "(Argentina) Bank Transfer deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "BANKTRANSFER_AR",
          "amount": 150,
          "currency": "ARS",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "documentType": "AR_DNI",
            "documentNumber": "33655208"
          }
        }
      },
      "BankTransferCrDepositRequest": {
        "summary": "(Costa Rica) Bank Transfer deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "BANKTRANSFER_CR",
          "amount": 150,
          "currency": "CRC",
          "customer": {
            "firstName": "John",
            "lastName": "Doe"
          }
        }
      },
      "BankTransferGtDepositRequest": {
        "summary": "(Guatemala) Bank Transfer deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "BANKTRANSFER_GT",
          "amount": 150,
          "currency": "GTM",
          "customer": {
            "firstName": "John",
            "lastName": "Doe"
          }
        }
      },
      "BankTransferHnDepositRequest": {
        "summary": "(Honduras) Bank Transfer deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "BANKTRANSFER_HN",
          "amount": 150,
          "currency": "HNL",
          "customer": {
            "firstName": "John",
            "lastName": "Doe"
          }
        }
      },
      "BankTransferPaDepositRequest": {
        "summary": "(Panama) Bank Transfer deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "BANKTRANSFER_PA",
          "amount": 150,
          "currency": "USD",
          "customer": {
            "firstName": "John",
            "lastName": "Doe"
          }
        }
      },
      "BankTransferUyDepositRequest": {
        "summary": "(Uruguay) Bank Transfer deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "BANKTRANSFER_UY",
          "amount": 150,
          "currency": "UYU",
          "customer": {
            "firstName": "John",
            "lastName": "Doe"
          }
        }
      },
      "CryptoBtcDepositRequest": {
        "summary": "Crypto (BTC) deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "CRYPTO",
          "amount": 1,
          "currency": "BTC",
          "customer": {
            "firstName": "John",
            "lastName": "Doe"
          }
        }
      },
      "CryptoBtcBep20DepositRequest": {
        "summary": "Crypto (BTC BEP-20) deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "CRYPTO",
          "amount": 1,
          "currency": "BTC",
          "customer": {
            "firstName": "John",
            "lastName": "Doe"
          },
          "additionalParameters": {
            "tokenContractStandard": "BEP-20"
          }
        }
      },
      "CryptoUsdtTrc20DepositRequest": {
        "summary": "Crypto (USDT TRC-20) deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "CRYPTO",
          "amount": 1,
          "currency": "USDT",
          "customer": {
            "firstName": "John",
            "lastName": "Doe"
          },
          "additionalParameters": {
            "tokenContractStandard": "TRC-20"
          }
        }
      },
      "FiatToCryptoDepositRequest": {
        "summary": "Crypto (EUR to USDT TRC-20) deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "CRYPTO",
          "amount": 1,
          "currency": "EUR",
          "customer": {
            "firstName": "John",
            "lastName": "Doe"
          },
          "additionalParameters": {
            "tokenContractStandard": "TRC-20",
            "cryptoCurrency": "USDT"
          }
        }
      },
      "PixDepositRequest": {
        "summary": "(Brazil) PIX deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "PIX",
          "amount": 100,
          "currency": "BRL",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "documentType": "BR_CPF",
            "documentNumber": "65745728000"
          }
        }
      },
      "BankTransferClDepositRequest": {
        "summary": "(Chile) Bank Transfer deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "BANKTRANSFER_CL",
          "amount": 1000,
          "currency": "CLP",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "documentType": "CL_CI",
            "documentNumber": "88682998"
          }
        }
      },
      "EfectyDepositRequest": {
        "summary": "(Colombia) EFECTY deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "EFECTY",
          "amount": 5000,
          "currency": "COP",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "phone": "57 3205671968",
            "documentType": "CO_CC",
            "documentNumber": "1045726853"
          }
        }
      },
      "NequiDepositRequest": {
        "summary": "(Colombia) NEQUI deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "NEQUI",
          "amount": 5000,
          "currency": "COP",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "phone": "57 3205671968",
            "documentType": "CO_CC",
            "documentNumber": "1045726853"
          }
        }
      },
      "PseDepositRequest": {
        "summary": "(Colombia) PSE deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "PSE",
          "amount": 5000,
          "currency": "COP",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "phone": "57 3205671968",
            "documentType": "CO_CC",
            "documentNumber": "1045726853"
          }
        }
      },
      "PaytmDepositRequest": {
        "summary": "(India) Paytm deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "PAYTM",
          "amount": 100,
          "currency": "INR",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "phone": "91 9876543210"
          }
        }
      },
      "PhonePeDepositRequest": {
        "summary": "(India) PhonePe deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "PHONEPE",
          "amount": 100,
          "currency": "INR",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "phone": "91 9876543210"
          }
        }
      },
      "UpiCollectDepositRequest": {
        "summary": "(India) UPI Collect deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "UPI",
          "amount": 100,
          "currency": "INR",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "phone": "91 9876543210",
            "accountNumber": "9876543210@upi"
          }
        }
      },
      "UpiIntentDepositRequest": {
        "summary": "(India) UPI Intent deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "UPI",
          "amount": 100,
          "currency": "INR",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "phone": "91 9876543210"
          }
        }
      },
      "SpeiDepositRequest": {
        "summary": "(Mexico) SPEI deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "SPEI",
          "amount": 220,
          "currency": "MXN",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "phone": "52 9616689020",
            "documentType": "MX_CURP",
            "documentNumber": "RICJ750331HDFVMN04"
          }
        }
      },
      "BankTransferPeDepositRequest": {
        "summary": "(Peru) Bank Transfer deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "BANKTRANSFER_PE",
          "amount": 100,
          "currency": "PEN",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "jd@mail.com",
            "documentType": "PE_DNI",
            "documentNumber": "123456789"
          }
        }
      },
      "PayPoDepositRequest": {
        "summary": "PayPo deposit",
        "value": {
          "paymentType": "DEPOSIT",
          "paymentMethod": "PAYPO",
          "amount": 150,
          "currency": "PLN",
          "customer": {
            "firstName": "John",
            "lastName": "Doe",
            "email": "my@email.com",
            "phone": "48 727516980"
          },
          "billingAddress": {
            "countryCode": "PL",
            "addressLine1": "Test",
            "city": "Warszawa",
            "postalCode": "02-672"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "The client sends HTTP requests with the Authorization header that contains the word Bearer followed by a space and the Shop Api Key.",
        "scheme": "bearer"
      },
      "SignatureAuth": {
        "type": "apiKey",
        "description": "HMAC-SHA256 hash generated from JSON body using Shop Signing Key as a secret.",
        "name": "Signature",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}